rfctr: move DescribeCubeSet into test_cube.py
Convention dictates all test classes for a module reside in the same
corresponding test module.
